requirements.txt
azure_storage==0.37.0
cachetools==5.3.2
Pillow==9.3.0
Pillow==10.2.0
Requests==2.31.0
websocket_client==1.6.4
websockets==12.0
